+++ src/test/java/com/ericsson/bss/cassandra/ecaudit/auth/TestAuditWhitelistManager.java
@@ -96,6 +96,13 @@ public class TestAuditWhitelistManager
         DatabaseDescriptor.clientInitialization(false);
     }
 
+    @Test
+    public void testSetupDelegation()
+    {
+        whitelistManager.setup();
+        verify(mockWhitelistDataAccess, times(1)).setup();
+    }
+
     @Test(expected = InvalidRequestException.class)
     public void testWhitelistAtCreateIsRejected()
     {