+++ CHANGES.md
@@ -1,11 +1,15 @@
 # Changes
 
+## Version 2.1.0
+* Make the audit whitelist table a protected resource in Cassandra
+
 ## Version 2.0.0
 * __NOTE__: This version is breaking backwards compatibility - consult detailed instructions in the [upgrade guide](UPGRADING.md)
 * Add support for whitelists based on specific operations
 * Make whitelist operations case insensitive
 * Remove support for whitelist management in CREATE ROLE statement
 * Limit whitelist management to one operation per statement
+* Backport ecAudit to Cassandra 3.0.11 (only in ecAudit for C* 3.0.11)
 
 ## Version 1.0.0
 * __NOTE__: This version is breaking backwards compatibility - consult detailed instructions in the [upgrade guide](UPGRADING.md)